requirements.txt
aiohttp==3.6.2
torch==1.6.0
torchvision==0.7.0
nltk==3.5
Pillow==7.2.0
tqdm==4.48.2
h5py==2.10.0
scipy==1.5.2
matplotlib==3.3.1
scikit-image==0.17.2
gensim==3.8.3
